US14 test created for desktop options
